DQ rule tests - consolidate v3 and deprecate stale v2
